Added health check component, axios, and status indicator
